% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_networks}
\alias{compare_networks}
\title{COmpare statistics from two simulated networks}
\usage{
compare_networks(net1, net2)
}
\arguments{
\item{net1}{a NetSim object}

\item{net2}{a NetSim object}
}
\value{
A dataframe
}
\description{
Generates summary statistics for two simulated networks of class
NetSim.
}
\details{
This function enables a quick comparison of two simulated networks
of class NetSim. This allows the user to see how different a network
simulated from a Halton sequence is from network simulated from a Poisson
point process with the same number of points in the same geographical window.
Similarly, it enables comparison of networks generated from different
spatial interaction functions. The density, mean degree, mean closeness, mean
betweenness, and the size of the largest connected component are computed for
each network.
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
