% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{HaltonSeq}
\alias{HaltonSeq}
\title{Simulate a 2D Halton sequence in a spatial window}
\usage{
HaltonSeq(points, window, seed)
}
\arguments{
\item{points}{the maximum number of points to simulate}

\item{window}{a window object of class spacejamr to use as the spatial extent}

\item{seed}{an optional seed}
}
\value{
An object of class PointSim that contains a geographical window of
class 'owin'. Within this window are four objects. n: the number of simulated
points. x: the x coordinates of the simulated points. y: the y coordinates of
the simulated points. markformat: an empty place holder.
}
\description{
Creates a new Halton Sequence in a spacejamr object
}
\details{
Note that the number of generated points may be less that the points
argument. If the spacejamr window is a rectangle, this function will generate
the specificed number of point. Otherwise, the specified number of points
will be generated within a containing rectangle and only those points within
the spacejamr window will be kept.
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
