% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexico.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Geographical administrative level 0 boundary of Mexico.}
\format{
A spacejamr object with an owin window object and an integer denoting
the projected coordinate reference system:
\describe{
\item{window}{geographical boundary of Mexico}
\item{crs}{the EPSG code for the coordinate reference system of the window}
...
}
}
\source{
\url{https://data.humdata.org/dataset/mexican-administrative-level-0-country-1-estado-and-2-municipio-boundary-polygons}
}
\usage{
mexico
}
\description{
A dataset containing Mexico's border and the coordinate reference system in
which Mexico's border is projected.
}
\keyword{datasets}
