% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{PowerLawNetwork}
\alias{PowerLawNetwork}
\title{Simulate a power law network}
\usage{
PowerLawNetwork(point_sim, base_prob, scale, threshold, power)
}
\arguments{
\item{point_sim}{a PointSim object}

\item{base_prob}{the theoretical probability that two nodes (points) with
distance 0 share a tie. Default is 0.9.}

\item{scale}{a coefficient to multiply the distance by. Default is 1.}

\item{threshold}{if two node exceed this probability, they will be coded as
having a tie. Default is 0.5.}

\item{power}{the exponent at which tie probability decays. Default is -2.8.}
}
\value{
An iGraph object
}
\description{
Simulates a spatial Bernoulli network from a NetSim object
using a standard power law function as the spatial interaction function.
}
\details{
The algorithm proceeds in three steps. First, it calculates the
distance between simulated points from a PointSim object, which includes
PointProcess and HaltonSeq objects. Then it calculates the distance between
all pairs of points. Finally, it uses a standard power law function to
calculate that any two point share a tie. If the threshold is exceeded, a tie
is created.
}
\examples{
\dontrun{
   # Create spacejamr object
   mex <- spacejamr("Z:shapefiles", "mexico_adm0")

   # With PointProcess
   mex_points <- PointProcess(1000, mex, 42)
   power_law <- PowerLawNetwork(mex_points, base_prob = 0.92, scale = 1,
                             threshold = 0.5, power = -2.4)

   # With HaltonSeq
   mex_seq <- HaltonSeq(1000, mex, 42)
   power_law <- PowerLawNetwork(mex_seq, base_prob = 0.98, scale = 100,
                             threshold = 0.5, power = -1.87)
}

}
\references{
Butts, Carter T. Spatial models of large-scale interpersonal
networks. Dissertation. (2002).
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
