% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{print.NetSim}
\alias{print.NetSim}
\title{Print information from a NetSim object}
\usage{
print(x)
}
\arguments{
\item{x}{a NetSim object}

\item{...}{ignored.}
}
\value{
No return value, called for sied effects
}
\description{
Plots a NetSim object and returns a ggraph object
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
