% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacejamr.R
\name{print.spacejamr}
\alias{print.spacejamr}
\title{Print information from a spacejamr instance}
\usage{
\method{print}{spacejamr}(x, ...)
}
\arguments{
\item{x}{a spacejamr object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Print method for the spacejamr class
}
\details{
Provides a wrapper for the print.owin method in the spatstats.geom
package.
}
\examples{
# Load spacejamr object
data("RI")

print(RI)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
