% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/spacejamr.R
\docType{package}
\name{spacejamr}
\alias{spacejamr}
\title{spacejamr: A package for simulating spatial Bernoulli networks}
\usage{
spacejamr(path)
}
\arguments{
\item{path}{the path to a shapefile as a string.}
}
\value{
a spacejamr object
}
\description{
The spacejamr package provides three categories of functions:
spacejamr, PointSim, and NetSim. They create new objects used in subsuquent
steps of network simulation, simulate spatial point processes, and create
network objects

Creates a new spacejamr object that for further analysis
}
\details{
The returned spacejamr object will contain a window object
containing a geographical boundary and its coordinate reference system.
Since simulating a spatial point process or sequnce requires a projected
coordinate reference system, this method will automatically look for the
most appropriate projected coordinate reference system and project the
geographical boundary to that system. Therefore, this function may take
some time to run but is necessary for later steps to simulate a spatial
Bernoulli network.
}
\section{spacejamr funnctions}{

APLNetwork
HaltonSeq
mexico
plot.NetSim
plot.PointSim
plot.spacejamr
PointProcess
PowerLawNetwork
print.NetSim
print.PointSim
print.spacejamr
spacejamr
summary.NetSim
summary.spacejamr
summary.PointSim
}

\examples{
\dontrun{
   mex <- spacejamr("Z:shapefiles/mexico_adm0.shp")
}

}
\author{
Darren Colby \cr
Email: dscolby@gmail.com
}
