% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{summary.NetSim}
\alias{summary.NetSim}
\title{Summary of NetSim graphs}
\usage{
summary(object)
}
\arguments{
\item{object}{a NetSim object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Prints a summary of a NetSim object
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
