% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{summary.PointSim}
\alias{summary.PointSim}
\title{Display summary information from a PointSim instance}
\usage{
\method{summary}{PointSim}(object, ...)
}
\arguments{
\item{object}{a PointSim object}

\item{...}{ignored.}
}
\value{
No return value
}
\description{
Prints a summary of information from either a PointProcess or
HaltonSeq object, whcih are both child classes of the PointSim class.
}
\examples{
# Load spacejamr object
data("RI")

# With PointProcess
ri_points <- PointProcess(10, RI, 42)
summary(ri_points)

# With HaltonSeq
ri_seq <- HaltonSeq(10, RI, 42)
summary(ri_seq)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
