% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacejamr.R
\name{summary.spacejamr}
\alias{summary.spacejamr}
\title{Print a summary of a spacejamr instance}
\usage{
\method{summary}{spacejamr}(object, ...)
}
\arguments{
\item{object}{a spacejamr object}

\item{...}{ignored.}
}
\description{
Summary method for the spacejamr class
}
\details{
Provides a wrapper for the summary.owin method in the spatstats.geom
package.
}
\examples{
\dontrun{
   mex <- spacejamr("Z:shapefiles", "mexico_adm0")
   summary(mex)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
