% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{plot.NetSim}
\alias{plot.NetSim}
\title{Plot a simulated network from a NetSim object}
\usage{
plot(x)
}
\arguments{
\item{x}{a NetSim graph}

\item{y}{ignored.}

\item{...}{ignored.}

\item{layout}{a layout to display the graph. Layout must be a valid string.
from the ggraph package. Default is "stress".}

\item{title}{an optional title.}

\item{node_color}{a color for the nodes. Default is blue.}

\item{edge_color}{a color for the edges. Default is red.}
}
\value{
A plot of classes 'ggraph' 'gg' and 'ggplot'
}
\description{
This can take either a PowerLawNetwork or APLNetwork object as
input, both of which are chidren of the NetSim class.
}
\details{
This method returns a ggraph object, which can be further refined
using standard ggraph and ggplot facilities.
}
\seealso{
\code{\link[ggraph:ggraph]{ggraph::ggraph()}}
}
\author{
Darren Colby \cr
Email:dscolby17@gmail.com
}
