% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacejamr.R
\name{plot.spacejamr}
\alias{plot.spacejamr}
\title{Plot the spatial extent of a spacejamr object}
\usage{
\method{plot}{spacejamr}(x, y, ..., title = "Spatial Window", fill = "blue")
}
\arguments{
\item{x}{an object of class spacejamr.}

\item{y}{ignored.}

\item{...}{ignored.}

\item{title}{an optional title for the plot. Default is "Spatial Window".}

\item{fill}{an optional fill for the plot. Default is blue.}
}
\value{
A plot of classes 'gg' and 'ggplot'
}
\description{
Plot method for the spacejamr class
}
\details{
The returned plot can be refined with standard ggplot2 functions
}
\examples{
# Load spacejamr object
data("RI")

plot(RI)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
