% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{PointProcess}
\alias{PointProcess}
\title{Simulate a spatial Poisson point process}
\usage{
PointProcess(points, window, seed)
}
\arguments{
\item{points}{the number of points to simulate}

\item{window}{a spacejamr object to use as the spatial extent}

\item{seed}{an optional seed}
}
\value{
An object of classes PointProcess and PointSim
}
\description{
Creates a new Poisson Point Process in a spacejamr object
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
