% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{PointProcess}
\alias{PointProcess}
\title{Simulate a spatial Poisson point process}
\usage{
PointProcess(points, window, seed)
}
\arguments{
\item{points}{the number of points to simulate}

\item{window}{a window object of class spacejamr to use as the spatial extent}

\item{seed}{an optional seed}
}
\value{
An object of classes PointProcess and PointSim
}
\description{
Creates a new Poisson Point Process in a spacejamr object
}
\examples{
\dontrun{
   # Create spacejamr object
   mex <- spacejamr("Z:shapefiles", "mexico_adm0")
   mex_points <- PointProcess(1000, mex, 42)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
