% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{PowerLawNetwork}
\alias{PowerLawNetwork}
\title{Simulate a power law network}
\usage{
PowerLawNetwork(point_sim, base_prob, scale, threshold, power)
}
\arguments{
\item{point_sim}{a PointSim object}

\item{base_prob}{the theoretical probability that two nodes (points) with
distance 0 share a tie. Default is 0.9.}

\item{scale}{a coefficient to multiply the distance by. Default is 1.}

\item{threshold}{if two node exceed this probability, they will be coded as
having a tie. Default is 0.5.}

\item{power}{the exponent at which tie probability decays. Default is -2.8.}
}
\value{
A network object of class 'igraph' that can be manipulated using the
'igraph' package.
}
\description{
Simulates a spatial Bernoulli network from a NetSim object
using a standard power law function as the spatial interaction function.
}
\details{
The algorithm proceeds in three steps. First, it calculates the
distance between simulated points from a PointSim object. Then it calculates
the distance between all pairs of points. Finally, it uses a standard power
law function to calculate that any two point share a tie. If the threshold is
exceeded, a tie is created.
}
\references{
Butts, Carter T. Spatial models of large-scale interpersonal
networks. Dissertation. (2002).
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
