% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{print.NetSim}
\alias{print.NetSim}
\title{Print information from a NetSim object}
\usage{
\method{print}{NetSim}(x, ...)
}
\arguments{
\item{x}{a NetSim object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Plots a NetSim object and returns a ggraph object
}
\examples{
# Create spacejamr object
data("RI")

ri_points <- PointSim(points = 10, window = RI, seed = 42)
spl_points <- NetSim(ri_points, base_prob = 0.92, scale = 1, threshold = 0.5,
                     power = -2.4)
print(spl_points)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
