% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{print.NetSim}
\alias{print.NetSim}
\title{Print information from a NetSim object}
\usage{
\method{print}{NetSim}(x, ...)
}
\arguments{
\item{x}{a NetSim object}

\item{...}{ignored.}
}
\value{
A ggraph object
}
\description{
Plots a NetSim object and returns a ggraph object
}
\examples{
\dontrun{
   # Create spacejamr object
   mex <- spacejamr("Z:shapefiles", "mexico_adm0")

   # With PointProcess
   mex_points <- PointProcess(1000, mex, 42)
   apl_points <- APLNetwork(mex_points, base_prob = 0.92, scale = 1,
                         threshold = 0.5, power = -2.4)
   print(apl_points)

   # With HaltonSeq
   mex_seq <- HaltonSeq(1000, mex, 42)
   apl_seq <- APLNetwork(mex_seq, base_prob = 0.98, scale = 100,
                      threshold = 0.5, power = -1.87)
   print(apl_seq)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
