% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{print.NetSim}
\alias{print.NetSim}
\title{Print information from a NetSim object}
\usage{
\method{print}{NetSim}(x, ...)
}
\arguments{
\item{x}{a NetSim object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Plots a NetSim object and returns a ggraph object
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
