% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{print.PointSim}
\alias{print.PointSim}
\title{Print information from a PointSim class}
\usage{
\method{print}{PointSim}(x, ...)
}
\arguments{
\item{x}{a PointSim object or a child object}

\item{...}{ignored.}
}
\description{
Print method for both the PointProcess and HaltonSeq classes,
which inherit methods from the PointSim class.
}
\examples{
\dontrun{
# Create spacejamr object
mex <- as.spacejamr("Z:shapefiles/mexico_adm0.shp")
}

\dontrun{
# With PointProcess
mex_points <- PointProcess(1000, mex, 42)
print(mex_points)
}

\dontrun{
# With HaltonSeq
mex_seq <- HaltonSeq(1000, mex, 42)
print(mex_seq)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
