% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{print.PointSim}
\alias{print.PointSim}
\title{Print information from a PointSim class}
\usage{
\method{print}{PointSim}(x, ...)
}
\arguments{
\item{x}{a PointSim object or a child object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Print method for both the PointProcess and HaltonSeq classes,
which inherit methods from the PointSim class.
}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
