% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{summary.NetSim}
\alias{summary.NetSim}
\title{Summary of NetSim graphs}
\usage{
\method{summary}{NetSim}(object, ...)
}
\arguments{
\item{object}{a NetSim object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Prints a summary of a NetSim object
}
\examples{
# Load spacejamr object
data("RI")

ri_points <- PointSim(points = 10, window = RI, seed = 42)
spl_points <- NetSim(ri_points, base_prob = 0.92, scale = 1, threshold = 0.5,
                     power = -2.4)
summary(spl_points)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
