% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{summary.NetSim}
\alias{summary.NetSim}
\title{Summary of NetSim graphs}
\usage{
\method{summary}{NetSim}(object, ...)
}
\arguments{
\item{object}{a NetSim object}

\item{...}{ignored.}
}
\value{
A ggraph object
}
\description{
Prints a summary of a NetSim object
}
\examples{
\dontrun{
# Create spacejamr object
mex <- as.spacejamr("Z:shapefiles/mexico_adm0.shp")
}

\dontrun{
# With PointProcess
mex_points <- PointProcess(1000, mex, 42)
apl_points <- APLNetwork(mex_points, base_prob = 0.92, scale = 1,
                         threshold = 0.5, power = -2.4)
summary(apl_points)
}

\dontrun{
# With HaltonSeq
mex_seq <- HaltonSeq(1000, mex, 42)
apl_seq <- APLNetwork(mex_seq, base_prob = 0.98, scale = 100,
                      threshold = 0.5, power = -1.87)
summary(apl_seq)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
