% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSim.R
\name{summary.NetSim}
\alias{summary.NetSim}
\title{Summary of NetSim graphs}
\usage{
\method{summary}{NetSim}(object, ...)
}
\arguments{
\item{object}{a NetSim object}

\item{...}{ignored.}
}
\value{
A ggraph object
}
\description{
Prints a summary of a NetSim object
}
\examples{
# Load spacejamr object
data("RI")

# With PointProcess
ri_points <- PointProcess(10, RI, 42)
apl_points <- APLNetwork(ri_points, base_prob = 0.92, scale = 1,
                         threshold = 0.5, power = -2.4)
summary(apl_points)

# With HaltonSeq
ri_seq <- HaltonSeq(1000, RI, 42)
apl_seq <- APLNetwork(ri_seq, base_prob = 0.98, scale = 100,
                      threshold = 0.5, power = -1.87)
summary(apl_seq)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
