% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{summary.PointSim}
\alias{summary.PointSim}
\title{Display summary information from a PointSim instance}
\usage{
\method{summary}{PointSim}(object, ...)
}
\arguments{
\item{object}{a PointSim object}

\item{...}{ignored.}
}
\description{
Prints a summary of information from either a PointProcess or
HaltonSeq object, whcih are both child classes of the PointSim class.
}
\examples{
\dontrun{
# Create spacejamr object
mex <- as.spacejamr("Z:shapefiles/mexico_adm0.shp")
}

\dontrun{
# With PointProcess
mex_points <- PointProcess(1000, mex, 42)
summary(mex_points)
}

\dontrun{
# With HaltonSeq
mex_seq <- HaltonSeq(1000, mex, 42)
summary(mex_seq)
}

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
