% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSim.R
\name{summary.PointSim}
\alias{summary.PointSim}
\title{Display summary information from a PointSim instance}
\usage{
\method{summary}{PointSim}(object, ...)
}
\arguments{
\item{object}{a PointSim object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Prints a summary of information from either a PointProcess or
HaltonSeq object, whcih are both child classes of the PointSim class.
}
\examples{
# Load spacejamr object
data("RI")

# With Poisson Process
ri_points <- PointSim(points = 10, window = RI, seed = 42)
summary(ri_points)

# With Halton Sequence
ri_seq <- HaltonSeq(points = 10, window = RI, type = halton, seed = 42)
summary(ri_seq)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
