% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacejamr.R
\name{summary.spacejamr}
\alias{summary.spacejamr}
\title{Print summary information of a spacejamr instance}
\usage{
\method{summary}{spacejamr}(object, ...)
}
\arguments{
\item{object}{a spacejamr object}

\item{...}{ignored.}
}
\value{
No return value, called for side effects
}
\description{
Summary method for the spacejamr class
}
\details{
Provides a wrapper for the summary.owin method in the spatstats.geom
package.
}
\examples{
# Load spacejamr object
data("RI")

summary(RI)

}
\author{
Darren Colby \cr
Email: dscolby17@gmail.com
}
